backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
paramiko>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
websockets>=11.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
httpx>=0.24.0
